requirements.txt
altgraph==0.17
future==0.18.2
numpy==1.18.1
pefile==2019.4.18
Pillow==7.0.0
PyInstaller==3.6
PyOpenGL==3.1.5
PySide2==5.14.1
pywin32-ctypes==0.2.0
shiboken2==5.14.1
